Refactor and trigger dynamic downloading at runtime
