+++ pkg/kubernetes/kubernetes.go
@@ -3,22 +3,15 @@ package kubernetes
 
 import (
 	"bufio"
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net/http"
-	"os"
-	"path/filepath"
 	"strings"
-	"time"
 
 	"get.porter.sh/porter/pkg/context"
 	"github.com/ghodss/yaml"
 	"github.com/gobuffalo/packr/v2"
 	"github.com/pkg/errors"
-	"github.com/rogpeppe/go-internal/semver"
 	"github.com/xeipuuv/gojsonschema"
 )
 
@@ -28,65 +21,22 @@ const (
 
 type Mixin struct {
 	*context.Context
-	schemas                 *packr.Box
+	schemas *packr.Box
+	KubectlDownloader
 	KubernetesClientVersion string
-}
-
-type KubectlVersion struct {
-	ClientVersion struct {
-		Major        string    `json:"major"`
-		Minor        string    `json:"minor"`
-		GitVersion   string    `json:"gitVersion"`
-		GitCommit    string    `json:"gitCommit"`
-		GitTreeState string    `json:"gitTreeState"`
-		BuildDate    time.Time `json:"buildDate"`
-		GoVersion    string    `json:"goVersion"`
-		Compiler     string    `json:"compiler"`
-		Platform     string    `json:"platform"`
-	} `json:"clientVersion"`
-	ServerVersion struct {
-		Major        string    `json:"major"`
-		Minor        string    `json:"minor"`
-		GitVersion   string    `json:"gitVersion"`
-		GitCommit    string    `json:"gitCommit"`
-		GitTreeState string    `json:"gitTreeState"`
-		BuildDate    time.Time `json:"buildDate"`
-		GoVersion    string    `json:"goVersion"`
-		Compiler     string    `json:"compiler"`
-		Platform     string    `json:"platform"`
-	} `json:"serverVersion"`
 }
 
 func New() *Mixin {
 	return &Mixin{
 		Context:                 context.New(),
 		schemas:                 NewSchemaBox(),
+		KubectlDownloader:       KubectlDownloaderImplementation{},
 		KubernetesClientVersion: defaultKubernetesClientVersion,
 	}
 }
 
 func NewSchemaBox() *packr.Box {
 	return packr.New("get.porter.sh/porter/pkg/kubernetes/schema", "./schema")
-}
-
-func (m *Mixin) reconcileKubectlVersion() error {
-
-	serverVersion, err := getKubectlServerVersion(m)
-
-	if err != nil {
-		return err
-	}
-	// install a new client if the current clientversion is null or older then the server version
-	if m.KubernetesClientVersion == "" || semver.Compare(m.KubernetesClientVersion, serverVersion) == -1 {
-		fmt.Fprintf(m.Out, "Kubectl server version (%s) does not match client version (%s); downloading a compatible client.\n",
-			serverVersion, m.KubernetesClientVersion)
-
-		err := installKubectlClient(m, serverVersion)
-		if err != nil {
-			return errors.Wrap(err, "unable to install a compatible kubectl client")
-		}
-	}
-	return err
 }
 
 func (m *Mixin) getCommandFile(commandFile string, w io.Writer) ([]byte, error) {
@@ -182,66 +132,4 @@ func (m *Mixin) handleOutputs(outputs []KubernetesOutput) error {
 		}
 	}
 	return nil
-}
-
-func installKubectlClient(m *Mixin, version string) error {
-
-	url := fmt.Sprintf("https://storage.googleapis.com/kubernetes-release/release/%s/bin/linux/amd64/kubectl", version)
-
-	// Fetch archive from url
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrap(err, "failed to construct GET request for fetching kubectl client binary")
-	}
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "failed to download kubectl client binary via url: %s", url)
-	}
-	defer res.Body.Close()
-
-	// Create a temp dir
-	tmpDir, err := m.FileSystem.TempDir("", "tmp")
-	if err != nil {
-		return errors.Wrap(err, "unable to create a temporary directory for downloading the kubectl client binary")
-	}
-	defer os.RemoveAll(tmpDir)
-
-	// Create the local binary
-	kubectlBinPath, err := m.FileSystem.Create(filepath.Join(tmpDir, "kubectlBin"))
-	if err != nil {
-		return errors.Wrap(err, "unable to create a local file for the kubectl client binary")
-	}
-
-	// Copy response body to local binary
-	_, err = io.Copy(kubectlBinPath, res.Body)
-	if err != nil {
-		return errors.Wrap(err, "unable to copy the kubectl client binary to the local archive file")
-	}
-
-	// Move the kubectl binary into the appropriate location
-	binPath := "/usr/local/bin/kubectl"
-	err = m.FileSystem.Rename(fmt.Sprintf("%s", kubectlBinPath.Name()), binPath)
-	if err != nil {
-		return errors.Wrapf(err, "unable to install the kubectl client binary to %q", binPath)
-	}
-	return nil
-}
-
-func getKubectlServerVersion(m *Mixin) (string, error) {
-
-	var stderr bytes.Buffer
-	currentKubectl := KubectlVersion{}
-
-	cmd := m.NewCommand("kubectl", "version", "-o", "json")
-	cmd.Stderr = &stderr
-
-	outputBytes, err := cmd.Output()
-	if err != nil {
-		return "", errors.Wrapf(err, "unable to determine kubernetes server version: %s", stderr.String())
-	}
-	// Rebuild version json object
-	json.Unmarshal(outputBytes, &currentKubectl)
-
-	version := currentKubectl.ServerVersion.GitVersion
-	return version, nil
 }
